requirements.txt
selenium==4.15.2
robotframework==6.1.1
robotframework-seleniumlibrary==6.2.0
robotframework-retryfailed==0.2.0
robotframework-pabot==2.16.0
openpyxl==3.1.2
allure-robotframework==2.13.2
testcontainers~=4.7.0